可以正常计算角度

+++ new-posys/Action_User/math/quarternion.h
@@ -3,10 +3,10 @@
 
 #include "config.h"
 
-void QuaternionInt(double quaternion[4],double data[3] );
-void QuaternionInt1(double quaternion[4],double data[3] );
-void Quaternion_to_Euler(const double quaternion[4],double Rad[3] );
-void Euler_to_Quaternion(const double Rad[3],double quaternion[4]);
+void QuaternionInt(long double quaternion[4],long double data[3] );
+void QuaternionInt1(long double quaternion[4],long double data[3] );
+void Quaternion_to_Euler(const long double quaternion[4],long double Rad[3] );
+void Euler_to_Quaternion(const long double Rad[3],long double quaternion[4]);
 #endif
